+++ intake/__init__.py
@@ -5,87 +5,81 @@
 # The full license is in the LICENSE file, distributed with this software.
 #-----------------------------------------------------------------------------
 
+import importlib
 import re
 import logging
 import warnings
-import sys
 
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
-from . import source
-from .source.base import Schema, DataSource
-from .catalog.base import RemoteCatalog
-from .catalog import local
-from .catalog.default import load_combo_catalog
-from .catalog.local import MergedCatalog, EntrypointsCatalog
-from .catalog.zarr import ZarrGroupCatalog
-from .container import upload
-from .source import register_driver, registry
-from .source.discovery import autodiscover
-from .gui import InstanceMaker
 
-# Populate list of autodetected drivers (plugins).
-# The built-in ones are included here because intake itself declares them via
-# its own entry_points.
-for name, driver in autodiscover().items():
-    register_driver(name, driver)
+from .source import register_driver, registry
+
+imports = {
+    "DataSource": "intake.source.base:DataSoruce",
+    'Schema': "intake.source.base:Schema",
+    "cat": "intake.catalog:builtin",
+    "upload": "intake.container:upload",
+    "gui": "intake.gui:instance",
+    "output_notebook": "intake.gui:output_notebook",
+    "register_driver": "intake.source:register_driver",
+    "unregister_driver": "intake.source:unregister_driver",
+}
+openers = {}
 
 logger = logging.getLogger('intake')
-if sys.version_info >= (3, 7):
-    def __getattr__(attr):
-        if attr == 'Catalog':
-            from .catalog.base import Catalog
-            warnings.warn('deprecation: intake.Catalog now references the base class intake.catalog.base.Catalog\n '
-                          'If you want to open a generic URL, you should use intake.open_catalog')
-            return Catalog
+
+
+def __getattr__(attr):
+    gl = globals()
+    if attr == 'Catalog':
+        from .catalog.base import Catalog
+        warnings.warn('deprecation: intake.Catalog now references the base class intake.catalog.base.Catalog\n '
+                      'If you want to open a generic URL, you should use intake.open_catalog')
+        return Catalog
+    elif attr in openers:
+        driver = openers[attr].load()
+        gl[attr] = driver
+    else:
+        if attr in gl:
+            return gl[attr]
+        elif attr in imports:
+            dest = imports[attr]
+            modname = dest.split(":", 1)[0]
+            logger.debug("Importing: %s" % modname)
+            mod = importlib.import_module(modname)
+            if ":" in dest:
+                gl[attr] = getattr(mod, dest.split(":")[1])
+            else:
+                gl[attr] = mod
+    return gl[attr]
+
+
+def __dir__():
+    return sorted(list(globals()) + list(openers) + list(imports))
 
 
 def make_open_functions():
     """From the current state of ``registry``, create open_* functions"""
-    for plugin_name, plugin in registry.items():
-        try:
-            func_name = 'open_' + plugin_name
-            if not func_name.isidentifier():
-                # primitive name normalization
-                func_name = re.sub('[-=~^&|@+]', '_', func_name)
-            if func_name.isidentifier():
-                globals()[func_name] = plugin
-            else:
-                warnings.warn('Invalid Intake plugin name "%s" found.' %
-                              plugin_name)
-        except:
-            logger.warning("Creation of open function failed for %s"
-                           "" % plugin_name)
+    import entrypoints
 
+    for plugin in entrypoints.get_group_all('intake.drivers'):
 
-def output_notebook(inline=True, logo=False):
-    """
-    Load the notebook extension
-
-    Parameters
-    ----------
-    inline : boolean (optional)
-        Whether to inline JS code or load it from a CDN
-    logo : boolean (optional)
-        Whether to show the logo(s)
-    """
-    try:
-        import hvplot
-    except ImportError:
-        raise ImportError("The intake plotting API requires hvplot."
-                          "hvplot may be installed with:\n\n"
-                          "`conda install -c pyviz hvplot` or "
-                          "`pip install hvplot`.")
-    import holoviews as hv
-    return hv.extension('bokeh', inline=inline, logo=logo)
+        register_driver(plugin.name, plugin, True)
+        func_name = 'open_' + plugin.name
+        if not func_name.isidentifier():
+            # primitive name normalization
+            func_name = re.sub('[-=~^&|@+]', '_', func_name)
+        if func_name.isidentifier():
+            # stash name for dir() and later fetch
+            openers[func_name] = plugin
+        else:
+            warnings.warn('Invalid Intake plugin name "%s" found.' %
+                          plugin.name)
 
 
 make_open_functions()
-cat = MergedCatalog(
-    [EntrypointsCatalog(), load_combo_catalog()],
-    name='builtin',
-    description='Generated from data packages found on your intake search path')
 
 
 def open_catalog(uri=None, **kwargs):
@@ -160,6 +154,3 @@ def open_catalog(uri=None, **kwargs):
         raise ValueError('Unknown catalog driver (%s), supply one of: %s'
                          % (driver, list(sorted(registry))))
     return registry[driver](uri, **kwargs)
-
-
-gui = InstanceMaker()
