+++ intake/gui/__init__.py
@@ -6,6 +6,8 @@
 #-----------------------------------------------------------------------------
 from distutils.version import LooseVersion
 
+gl = globals()
+
 
 def do_import():
     error = too_old = False
@@ -16,56 +18,44 @@ def do_import():
         error = e
 
     if too_old or error:
-        class GUI(object):
-            def __repr__(self):
-                raise RuntimeError("Please install panel to use the GUI `conda "
-                                   "install -c conda-forge panel>=0.8.0`. Import "
-                                   "failed with error: %s" % error)
-        return GUI
+        raise RuntimeError("Please install panel to use the GUI `conda "
+                           "install -c conda-forge panel>=0.8.0`. Import "
+                           "failed with error: %s" % error)
 
+    from .gui import GUI
+    css = """
+    .scrolling {
+      overflow: scroll;
+    }
+    """
+    pn.config.raw_css.append(css)  # add scrolling class from css (panel GH#383, GH#384)
+    pn.extension()
+    gl['instance'] = GUI()
+
+
+def __getattr__(attr):
+    if attr == 'instance':
+        do_import()
+    return gl['instance']
+
+
+def output_notebook(inline=True, logo=False):
+    """
+    Load the notebook extension
+
+    Parameters
+    ----------
+    inline : boolean (optional)
+        Whether to inline JS code or load it from a CDN
+    logo : boolean (optional)
+        Whether to show the logo(s)
+    """
     try:
-        from .gui import GUI
-        css = """
-        .scrolling {
-          overflow: scroll;
-        }
-        """
-        pn.config.raw_css.append(css)  # add scrolling class from css (panel GH#383, GH#384)
-        pn.extension()
-
-    except Exception as e:
-
-        class GUI(object):
-            def __repr__(self):
-                raise RuntimeError("Initialisation of GUI failed, even though "
-                                   "panel is installed. Please update it "
-                                   "to a more recent version (`conda install -c"
-                                   " conda-forge panel>=0.9.5`).")
-    return GUI
-
-
-class InstanceMaker(object):
-
-    def __init__(self):
-        self._instance = None
-
-    def _instantiate(self):
-        if self._instance is None:
-            GUI = do_import()
-            self._instance = GUI()
-
-    def __getattr__(self, attr, *args, **kwargs):
-        self._instantiate()
-        return getattr(self._instance, attr, *args, **kwargs)
-
-    def __getitem__(self, item):
-        self._instantiate()
-        return self._instance[item]
-
-    def __repr__(self):
-        self._instantiate()
-        return repr(self._instance)
-
-    def __dir__(self):
-        self._instantiate()
-        return dir(self._instance)
+        import hvplot
+    except ImportError:
+        raise ImportError("The intake plotting API requires hvplot."
+                          "hvplot may be installed with:\n\n"
+                          "`conda install -c pyviz hvplot` or "
+                          "`pip install hvplot`.")
+    import holoviews as hv
+    return hv.extension('bokeh', inline=inline, logo=logo)
