+++ intake/source/__init__.py
@@ -5,13 +5,29 @@
 # The full license is in the LICENSE file, distributed with this software.
 #-----------------------------------------------------------------------------
 
+import entrypoints
 import logging
-logger = logging.getLogger('intake')
 
 from ..utils import DriverRegistryView
+from .base import DataSource
 
-# The registry is the mapping of plugin name (aka driver) to DataSource class
-_registry = {}  # internal mutable registry
+logger = logging.getLogger('intake')
+
+
+class DriverRegistry(dict):
+    """Dict of driver: DataSource class
+
+    If the value object is a EntryPoint, will load it when accesses, which
+    does the import.
+    """
+
+    def __getitem__(self, item):
+        if isinstance(super().__getitem__(item), entrypoints.EntryPoint):
+            self[item] = super().__getitem__(item).load()
+        return super().__getitem__(item)
+
+
+_registry = DriverRegistry()  # internal mutable registry
 registry = DriverRegistryView(_registry)  # public, read-ony wrapper
 
 
@@ -63,7 +79,7 @@ def unregister_driver(name):
     return _registry.pop(name, None)
 
 
-# A set of fully-qualifies package.module.Class mappings
+# A set of fully-qualified package.module.Class mappings
 classes = {}
